working on real quick sort pivot helper
after working on my own version of quick sort, I want to try to learn
how to do the real version. This first part is for the pivot helper
method. This commit is explicity for my own version of the pivot helper
method. I got it working, but after watching the vid on this part, I
think I'll have to change it slightly to keep the same approach the vid
is doing